utils: Move 'EggTimer' class to 'ost_utils.utils'
