+++ ost_utils/utils.py
@@ -11,8 +11,24 @@ import os
 import queue
 import sys
 import threading
+import time
 
 LOGGER = logging.getLogger(__name__)
+
+
+class EggTimer:
+    def __init__(self, timeout):
+        self.timeout = timeout
+
+    def __enter__(self):
+        self.start_time = time.time()
+        return self
+
+    def __exit__(self, *_):
+        pass
+
+    def elapsed(self):
+        return (time.time() - self.start_time) > self.timeout
 
 
 def _ret_via_queue(func, queue):
